move lazy out for merge
